+++ backend/controller/TourFAQController.js
@@ -19,6 +19,7 @@ export const FAQCreate = async (req, res) => {
 
 export const FAQDelete = async (req, res) => {
 	const { id } = req.params
+	if(!id && id === undefined) return res.status(401).json({message:'id is invalid'})
 	const findFAQ = await prisma.t_faq.delete({
 		where: {
 			id: Number(id),
@@ -46,6 +47,7 @@ export const FAQShowAll = async (req, res) => {
 
 export const FAQShow = async (req, res) => {
 	const { id } = req.params
+	if(!id && id === undefined) return res.status(401).json({message:'id is invalid'})
 
 	const FAQ = await prisma.t_faq.findUnique({
 		where: {
@@ -61,6 +63,7 @@ export const FAQShow = async (req, res) => {
 
 export const FAQEdit = async (req, res) => {
 	const { id } = req.params
+	if(!id && id === undefined) return res.status(401).json({message:'id is invalid'})
 	const { name, description, archive } = req.body
 
 	const FAQ = await prisma.t_faq.update({
