+++ backend/controller/TourTypeController.js
@@ -59,6 +59,7 @@ export const tourShowAll = async (req, res) => {
 
 export const tourShow = async (req, res) => {
     const {id} = req.params
+    if(!id && id === undefined) return res.status(401).json({message:'id is invalid'})
 
     const tours = await prisma.t_types.findUnique({
         where: {
@@ -71,6 +72,7 @@ export const tourShow = async (req, res) => {
 
 export const tourTypeShowUrl = async (req, res) => {
     const {url} = req.params
+    if(!url && url === undefined) return res.status(401).json({message:'url is invalid'})
 
     const tourType = await prisma.t_types.findFirst({
         where:{
@@ -97,6 +99,7 @@ export const tourTypeShowUrl = async (req, res) => {
 
 export const tourEdit = async (req, res) => {
     const {id} = req.params
+    if(!id && id === undefined) return res.status(401).json({message:'id is invalid'})
     const {
         parent,
         name,
@@ -150,6 +153,7 @@ export const tourEdit = async (req, res) => {
 
 export const tourDelete = async (req, res) => {
     const {id} = req.params
+    if(!id && id === undefined) return res.status(401).json({message:'id is invalid'})
 
     const tour_delete = await prisma.t_types.delete({
         where: {
